chore: styled 폴더 분리
